docs: add Azure SWA as deployment option (#7407)
I've linked to the dev.to article (which is a timestamped stable version) - to be consistent with other deployment snippets.

The dev.to article references [this GitHub repo](https://github.com/fearlessly-dev/swa-demo-docusaurus) and  this[#30DaysOfSWA](https://www.azurestaticwebapps.dev/blog/build-with-docusaurus) site which was the inspiration for writing this article. The repo contains the code for the demo, and its README is the canonical source for the article, in case relevant.